CLN refactorize compat+clean unused
