+++ loky/backend/compat_posix.py
@@ -1,3 +1,4 @@
+# flake8: noqa
 ###############################################################################
 # Compat file to load the correct wait function
 #
@@ -5,10 +6,8 @@
 #
 import sys
 
+# Compat wait
 if sys.version_info < (3, 3):
-    if sys.platform == "win32":
-        from ._win_wait import wait
-    else:
-        from ._posix_wait import wait
+    from ._posix_wait import wait
 else:
-    from multiprocessing.connection import wait  # noqa: F401
+    from multiprocessing.connection import wait
