+++ loky/backend/process.py
@@ -8,10 +8,7 @@
 #
 import os
 import sys
-if sys.version_info >= (3, 4):
-    from multiprocessing.process import BaseProcess
-else:
-    from multiprocessing.process import Process as BaseProcess
+from .compat import BaseProcess
 
 
 class LokyProcess(BaseProcess):
