+++ loky/backend/spawn.py
@@ -165,7 +165,6 @@ def prepare(data):
     if 'init_main_from_name' in data:
         _fixup_main_from_name(data['init_main_from_name'])
     elif 'init_main_from_path' in data:
-        print(data['init_main_from_path'])
         _fixup_main_from_path(data['init_main_from_path'])
