+++ loky/process_executor.py
@@ -72,18 +72,12 @@ from functools import partial
 
 from . import _base
 from .backend import get_context
-from .backend.connection import wait
+from .backend.compat import queue
+from .backend.compat import wait, PicklingError
 from .backend.queues import Queue, SimpleQueue
 
 # Compatibility for python2.7
-if sys.version_info[:2] > (2, 7):
-    import queue
-    from queue import Empty
-    from _pickle import PicklingError
-else:
-    import Queue as queue
-    from Queue import Empty
-    from pickle import PicklingError
+if sys.version_info[0] == 2:
     ProcessLookupError = OSError
 
 
@@ -329,7 +323,7 @@ def _process_worker(call_queue, result_queue, processes_management_lock,
             call_item = call_queue.get(block=True, timeout=timeout)
             if call_item is None:
                 mp.util.info("shutting down worker on sentinel")
-        except Empty:
+        except queue.Empty:
             mp.util.info("shutting down worker after timeout %0.3fs"
                          % timeout)
             if processes_management_lock.acquire(block=False):
@@ -388,7 +382,7 @@ def _add_call_item_to_queue(pending_work_items,
             return
         try:
             work_id = work_ids.get(block=False)
-        except Empty:
+        except queue.Empty:
             return
         else:
             work_item = pending_work_items[work_id]
@@ -483,7 +477,6 @@ def _queue_management_worker(executor_reference,
                                 call_queue)
         # Wait for a result to be ready in the result_queue while checking
         # that worker process are still running. If a worker process
-        count = 0
         while not wakeup.get_and_unset():
             worker_sentinels = [p.sentinel for p in processes.values()]
             ready = wait([result_reader] + worker_sentinels,
